3-puts.c

+++ 0x05-pointers_arrays_strings/3-puts.c
@@ -6,9 +6,9 @@
 */
 void _puts(char *str)
 {
-	int i;
+	long int i;
 
-	for (i = 0; str[i]; i ++)
+	for (i = 0; str[i]; i++)
 	_putchar(str[i]);
 	_putchar('\n');
 }
